+++ test/kt_iterable_test.dart
@@ -3,113 +3,200 @@ import 'package:ktc_dart/ktc_dart.dart';
 import 'package:test/test.dart';
 
 void main() {
+  final Iterable<int> empty = Iterable.empty();
   late Iterable<int> iterable;
+  late Iterable<int> reversedIterable;
 
   setUp(() {
     iterable = Iterable.generate(3);
+    reversedIterable = Iterable.generate(
+      3,
+      (index) => 2 - index,
+    );
   });
 
   test('average', () {
-    expect(<num>[].average, 0.0);
-    expect([1, 2].average, 1.5);
-    expect([1.0, 2.0].average, 1.5);
-    expect([double.nan, double.nan].average, isNaN);
-    expect([double.nan, 1.1].average, isNaN);
-    expect([double.nan, 1].average, isNaN);
-    expect([1, 2.0].average, 1.5);
-    expect([0, 0].average, 0.0);
+    expect(empty.average, 0.0);
+    expect(Iterable.generate(2, (index) => index + 1).average, 1.5);
+    expect(
+      Iterable.generate(2, (index) => (index + 1).toDouble()).average,
+      1.5,
+    );
+    expect(Iterable.generate(2, (index) => double.nan).average, isNaN);
+    expect(Iterable.generate(2, (index) => 0 / index).average, isNaN);
+    expect(
+      Iterable.generate(2, (index) => (0 / index).toDouble()).average,
+      isNaN,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? index.toDouble() : index,
+      ).average,
+      0.5,
+    );
+    expect(Iterable.generate(2, (index) => 0).average, 0.0);
 
-    expect([1, -2].average, -0.5);
-    expect([1.0, -2.0].average, -0.5);
-    expect([double.nan, -double.nan].average, isNaN);
-    expect([double.nan, -1.1].average, isNaN);
-    expect([double.nan, -1].average, isNaN);
-    expect([1, -2.0].average, -0.5);
-    expect([0, -0].average, 0.0);
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -index : index,
+      ).average,
+      -0.5,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -index : index,
+      ).average,
+      -0.5,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -double.nan : double.nan,
+      ).average,
+      isNaN,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -index.toDouble() : double.nan,
+      ).average,
+      isNaN,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -index : double.nan,
+      ).average,
+      isNaN,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -index.toDouble() : index,
+      ).average,
+      -0.5,
+    );
+    expect(
+      Iterable.generate(
+        2,
+        (index) => index.isOdd ? -0.0 : 0.0,
+      ).average,
+      0.0,
+    );
   });
 
   group('Associate', () {
     test('associate', () {
-      final associated = iterable.associate((element) => MapEntry(
+      MapEntry transform(int element) => MapEntry(
             element.toString(),
             element * 2,
-          ));
+          );
 
-      expect(associated, {'0': 0, '1': 2, '2': 4});
+      expect(empty.associate(transform), {});
+      expect(
+        iterable.associate(transform),
+        {'0': 0, '1': 2, '2': 4},
+      );
     });
 
     test('associateBy', () {
-      final associated = iterable.associateBy((element) => element.toString());
+      String keySelector(int element) => element.toString();
 
-      expect(associated, {'0': 0, '1': 1, '2': 2});
+      expect(empty.associateBy(keySelector), {});
+      expect(iterable.associateBy(keySelector), {'0': 0, '1': 1, '2': 2});
     });
 
     test('associateAndTransformBy', () {
-      final associated = iterable.associateAndTransformBy(
-        (element) => element.toString(),
-        (element) => element * 2,
-      );
+      String keySelector(int element) => element.toString();
+      int valueTransform(int element) => element * 2;
 
-      expect(associated, {'0': 0, '1': 2, '2': 4});
+      expect(empty.associateAndTransformBy(keySelector, valueTransform), {});
+      expect(
+        iterable.associateAndTransformBy(keySelector, valueTransform),
+        {'0': 0, '1': 2, '2': 4},
+      );
     });
 
     test('associaateWith', () {
-      final associated = iterable.associateWith(
-        (element) => element.toString(),
-      );
+      String valueSelector(int element) => element.toString();
 
-      expect(associated, {0: '0', 1: '1', 2: '2'});
+      expect(empty.associateWith(valueSelector), {});
+      expect(iterable.associateWith(valueSelector), {0: '0', 1: '1', 2: '2'});
+    });
+  });
+
+  group('Chunked', () {
+    test('chunked', () {
+      expect(() => empty.chunked(-1), throwsArgumentError);
+      expect(() => empty.chunked(0), throwsArgumentError);
+      expect(empty.chunked(1), []);
+      expect(() => iterable.chunked(-1), throwsArgumentError);
+      expect(() => iterable.chunked(0), throwsArgumentError);
+      expect(iterable.chunked(1), [
+        [0],
+        [1],
+        [2],
+      ]);
+      expect(iterable.chunked(2), [
+        [0, 1],
+        [2],
+      ]);
+      expect(iterable.chunked(3), [
+        [0, 1, 2]
+      ]);
+      expect(iterable.chunked(4), [
+        [0, 1, 2]
+      ]);
     });
 
-    group('Chunked', () {
-      test('chunked', () {
-        final chunked = iterable.chunked(2);
-
-        expect(chunked, [
-          [0, 1],
-          [2],
-        ]);
-      });
-
-      test('chunkedAndTransform', () {
-        final chunked = iterable.chunkedAndTransform(
+    test('chunkedAndTransform', () {
+      expect(empty, []);
+      expect(
+        iterable.chunkedAndTransform(
           2,
           (chunk) => chunk.reduce((value, element) => value + element),
-        );
-
-        expect(chunked, [1, 2]);
-      });
+        ),
+        [1, 2],
+      );
     });
   });
 
   test('count', () {
+    expect(empty.count(), 0);
     expect(iterable.count(), 3);
     expect(iterable.count((element) => element.isEven), 2);
   });
 
   group('Distinct', () {
     test('distinct', () {
-      final distinct = [...iterable, ...iterable].distinct;
-
-      expect(distinct, [0, 1, 2]);
+      expect(empty.distinct, []);
+      expect(Iterable.generate(2, (index) => 0).distinct, [0]);
     });
 
     test('distinctBy', () {
-      final distinctBy = iterable.distinctBy((element) => element.isEven);
+      bool selector(int element) => element.isEven;
 
-      expect(distinctBy, [0, 1]);
+      expect(empty.distinctBy(selector), []);
+      expect(iterable.distinctBy(selector), [0, 1]);
     });
   });
 
   group('ElementAt', () {
     test('ElementAtOrElse', () {
-      expect(iterable.elementAtOrElse(0, (index) => -index), 0);
-      expect(iterable.elementAtOrElse(2, (index) => -index), 2);
-      expect(iterable.elementAtOrElse(-1, (index) => -index), 1);
-      expect(iterable.elementAtOrElse(3, (index) => -index), -3);
+      int orElse(int index) => -index;
+
+      expect(empty.elementAtOrElse(0, orElse), -0);
+      expect(iterable.elementAtOrElse(0, orElse), 0);
+      expect(iterable.elementAtOrElse(2, orElse), 2);
+      expect(iterable.elementAtOrElse(-1, orElse), 1);
+      expect(iterable.elementAtOrElse(3, orElse), -3);
     });
 
     test('ElementAtOrNull', () {
+      expect(empty.elementAtOrNull(0), null);
       expect(iterable.elementAtOrNull(0), 0);
       expect(iterable.elementAtOrNull(2), 2);
       expect(iterable.elementAtOrNull(-1), null);
@@ -118,14 +205,18 @@ void main() {
   });
 
   test('expandIndexed', () {
-    final expandIndexed =
-        iterable.expandIndexed((index, element) => [index, element]);
+    Iterable<int> transform(int index, int element) => [index, element];
 
-    expect(expandIndexed, [0, 0, 1, 1, 2, 2]);
+    expect(empty.expandIndexed(transform), []);
+    expect(iterable.expandIndexed(transform), [0, 0, 1, 1, 2, 2]);
   });
 
   group('First', () {
     test('firstNotNullOf', () {
+      expect(
+        () => empty.firstNotNullOf((element) => element),
+        throwsA(isA<NoSuchElementException>()),
+      );
       expect(iterable.firstNotNullOf((element) => element.toString()), '0');
       expect(
         () => iterable.firstNotNullOf((element) => null),
@@ -134,6 +225,7 @@ void main() {
     });
 
     test('firstNotNullOfOrNull', () {
+      expect(empty.firstNotNullOfOrNull((element) => element), null);
       expect(
         iterable.firstNotNullOfOrNull((element) => element.toString()),
         '0',
@@ -145,61 +237,60 @@ void main() {
     });
 
     test('firstOrNull', () {
-      expect([].firstOrNull, null);
+      expect(empty.firstOrNull, null);
       expect(iterable.firstOrNull, 0);
     });
 
     test('firstWhereOrNull', () {
-      expect([].firstWhereOrNull((element) => element.isEven), null);
+      expect(empty.firstWhereOrNull((element) => element.isEven), null);
       expect(iterable.firstWhereOrNull((element) => element.isEven), 0);
       expect(iterable.firstWhereOrNull((element) => element.isNegative), null);
     });
   });
 
   test('flatten', () {
+    expect(Iterable.generate(1, (index) => Iterable.empty()).flatten, []);
     expect(
-      [
-        [1, 2],
-        [3, 4]
-      ].flatten,
-      [1, 2, 3, 4],
+      Iterable.generate(
+        2,
+        (i) => Iterable.generate(2, (j) => i + j),
+      ).flatten,
+      [0, 1, 1, 2],
     );
 
     expect(
-      [
-        [1, 2],
-        [
-          3,
-          4,
-          [5, 6]
-        ]
-      ].flatten,
-      [
-        1,
+      Iterable.generate(
         2,
-        3,
-        4,
-        [5, 6]
+        (index) => index == 0
+            ? Iterable.generate(2)
+            : Iterable.generate(
+                1,
+                (index) => Iterable.generate(2),
+              ),
+      ).flatten,
+      [
+        0,
+        1,
+        [0, 1]
       ],
     );
   });
 
   group('Fold', () {
     test('foldIndexed', () {
-      expect(
-        iterable.foldIndexed<int>(
-          0,
-          (index, previousValue, element) => index + previousValue + element,
-        ),
-        6,
-      );
+      int combine(int index, int previousValue, int element) =>
+          index + previousValue + element;
+
+      expect(empty.foldIndexed<int>(0, combine), 0);
+      expect(iterable.foldIndexed<int>(0, combine), 6);
     });
 
     test('runningFold', () {
       final initialValue = 'start';
-      String combine(previousValue, element) => '$previousValue$element';
+      String combine(String previousValue, int element) =>
+          '$previousValue$element';
 
-      expect([].runningFold(initialValue, combine), ['start']);
+      expect(empty.runningFold(initialValue, combine), ['start']);
       expect(
         iterable.runningFold(initialValue, combine),
         ['start', 'start0', 'start01', 'start012'],
@@ -208,10 +299,10 @@ void main() {
 
     test('runningFoldIndexed', () {
       final initialValue = 'start';
-      String combine(index, previousValue, element) =>
+      String combine(int index, String previousValue, int element) =>
           '$previousValue$element$index';
 
-      expect([].runningFoldIndexed(initialValue, combine), ['start']);
+      expect(empty.runningFoldIndexed(initialValue, combine), ['start']);
       expect(
         iterable.runningFoldIndexed(initialValue, combine),
         ['start', 'start00', 'start0011', 'start001122'],
@@ -222,31 +313,36 @@ void main() {
   test('forEachIndexed', () {
     var i = 0;
     var iterator = iterable.iterator..moveNext();
-
-    iterable.forEachIndexed((index, element) {
+    void action(int index, int element) {
       expect(index, i++);
       expect(index, iterator.current);
       iterator.moveNext();
-    });
+    }
+
+    empty.forEachIndexed(action);
+    iterable.forEachIndexed(action);
   });
 
   group('GroupBy', () {
     test('groupBy', () {
-      final groupBy = iterable.groupBy((element) => element.isEven);
+      bool keySelector(int element) => element.isEven;
 
-      expect(groupBy, {
+      expect(empty.groupBy(keySelector), {});
+      expect(iterable.groupBy(keySelector), {
         true: [0, 2],
         false: [1],
       });
     });
 
     test('groupAndTransformBy', () {
-      final grogroupAndTransformBy = iterable.groupAndTransformBy(
-        (element) => element.isEven,
-        (element) => element.toString(),
-      );
+      bool keySelector(int element) => element.isEven;
+      String valueTransform(int element) => element.toString();
 
-      expect(grogroupAndTransformBy, {
+      expect(
+        empty.groupAndTransformBy(keySelector, valueTransform),
+        {},
+      );
+      expect(iterable.groupAndTransformBy(keySelector, valueTransform), {
         true: ['0', '2'],
         false: ['1'],
       });
@@ -255,6 +351,7 @@ void main() {
 
   group('IndexOf', () {
     test('indexOf', () {
+      expect(empty.indexOf(0), -1);
       expect(iterable.indexOf(-1), -1);
       expect(iterable.indexOf(0), 0);
       expect(iterable.indexOf(2), 2);
@@ -262,18 +359,25 @@ void main() {
     });
 
     test('indexOfFirst', () {
+      expect(empty.indexOfFirst((element) => element.isNegative), -1);
       expect(iterable.indexOfFirst((element) => element.isNegative), -1);
       expect(iterable.indexOfFirst((element) => element.isEven), 0);
     });
 
     test('indexOfLast', () {
+      expect(empty.indexOfLast((element) => element.isNegative), -1);
       expect(iterable.indexOfLast((element) => element.isNegative), -1);
       expect(iterable.indexOfLast((element) => element.isEven), 2);
     });
 
     test('lastIndexOf', () {
-      iterable = [0, 0, 1, 1, 2, 2];
+      iterable = Iterable<Iterable<int>>.generate(
+        3,
+        (index) => Iterable<int>.generate(2, (_) => index),
+      ).flatten;
+      // [0, 0, 1, 1, 2, 2];
 
+      expect(empty.lastIndexOf(0), -1);
       expect(iterable.lastIndexOf(-1), -1);
       expect(iterable.lastIndexOf(0), 1);
       expect(iterable.lastIndexOf(2), 5);
@@ -282,56 +386,64 @@ void main() {
   });
 
   test('intersect', () {
-    expect(iterable.intersect([1, 2]), <int>{1, 2});
-    expect(iterable.intersect([2, 3]), <int>{2});
-    expect(iterable.intersect([3, 4]), <int>{});
-    expect(iterable.intersect([1, 1]), <int>{1});
-    expect([1, 2, 2, 3].intersect([1, 2]), <int>{1, 2});
+    expect(empty.intersect([1, 2]), <int>[]);
+    expect(iterable.intersect([1, 2]), <int>[1, 2]);
+    expect(iterable.intersect([2, 3]), <int>[2]);
+    expect(iterable.intersect([3, 4]), <int>[]);
+    expect(iterable.intersect([1, 1]), <int>[1]);
+    expect(
+      Iterable.generate(
+        6,
+        (index) => index.isOdd ? index - 1 : index,
+      ).intersect([1, 2, 3, 4]),
+      <int>[2, 4],
+    );
   });
 
   test('joinToString', () {
-    final iterable = [
-      'apple',
-      'banana',
-      'cherry',
-      'durian',
-      'fig',
-      'grape',
-      'honeydew',
-    ];
-    final result1 = iterable.joinToString();
-    final result2 = iterable.joinToString(limit: 3);
-    final result3 = iterable.joinToString(
-      prefix: 'fruits: ',
-      postfix: 'delicious',
-      separator: ' & ',
-      transform: (element) => element.toUpperCase(),
-      truncated: 'more... ',
-      limit: 3,
-    );
+    final iterable = Iterable.generate(6, (index) => index.toString());
 
+    expect(empty.joinToString(), '');
     expect(
-      result1,
-      'apple, banana, cherry, durian, fig, grape, honeydew',
+      iterable.joinToString(),
+      '0, 1, 2, 3, 4, 5',
+    );
+    expect(empty.joinToString(limit: 3), '');
+    expect(
+      iterable.joinToString(limit: 3),
+      '0, 1, 2, ...',
     );
     expect(
-      result2,
-      'apple, banana, cherry, ...',
-    );
+        empty.joinToString(
+          prefix: 'start ',
+          postfix: 'end',
+          separator: ' & ',
+          transform: (element) => '-$element',
+          truncated: 'more... ',
+          limit: 3,
+        ),
+        'start end');
     expect(
-      result3,
-      'fruits: APPLE & BANANA & CHERRY & more... delicious',
+      iterable.joinToString(
+        prefix: 'start ',
+        postfix: 'end',
+        separator: ' & ',
+        transform: (element) => '-$element',
+        truncated: 'more... ',
+        limit: 3,
+      ),
+      'start -0 & -1 & -2 & more... end',
     );
   });
 
   group('Last', () {
     test('lastOrNull', () {
-      expect([].lastOrNull, null);
+      expect(empty.lastOrNull, null);
       expect(iterable.lastOrNull, 2);
     });
 
     test('lastWhereOrNull', () {
-      expect([].lastWhereOrNull((element) => element.isEven), null);
+      expect(empty.lastWhereOrNull((element) => element.isEven), null);
       expect(iterable.lastWhereOrNull((element) => element.isEven), 2);
       expect(iterable.lastWhereOrNull((element) => element.isNegative), null);
     });
@@ -339,90 +451,89 @@ void main() {
 
   group('Map', () {
     test('mapIndexed', () {
-      final result = iterable.mapIndexed((index, element) => index + element);
+      int transform(int index, int element) => index + element;
 
-      expect(result, [0, 2, 4]);
+      expect(empty.mapIndexed(transform), []);
+      expect(iterable.mapIndexed(transform), [0, 2, 4]);
     });
 
     test('mapIndexedNotNull', () {
-      final result = iterable.mapIndexedNotNull(
-        (index, element) => (index * element).isEven ? null : element,
-      );
+      int? transform(int index, int element) =>
+          (index * element).isEven ? null : element;
 
-      expect(result, [1]);
+      expect(empty.mapIndexedNotNull(transform), []);
+      expect(iterable.mapIndexedNotNull(transform), [1]);
     });
 
     test('mapNotNull', () {
-      final result = iterable.mapNotNull(
-        (element) => element.isEven ? null : element,
-      );
+      int? transform(int element) => element.isEven ? null : element;
 
-      expect(result, [1]);
+      expect(empty.mapNotNull(transform), []);
+      expect(iterable.mapNotNull(transform), [1]);
     });
   });
 
   group('Max', () {
     test('maxOrNull', () {
-      expect(<num>[].maxOrNull, null);
+      expect(empty.maxOrNull, null);
       expect(iterable.maxOrNull, 2);
     });
 
     test('maxByOrNull', () {
-      final iterable = [
-        [1],
-        [1, 2],
-        [1, 2, 3],
-      ];
-      Comparable selector(List element) => element.length;
+      final iterable = Iterable.generate(
+        3,
+        (index) => Iterable<int>.generate(index + 1),
+      );
+      Comparable selector(Iterable<int> element) => element.length;
 
-      expect(<List>[].maxByOrNull(selector), null);
+      expect(Iterable<Iterable<int>>.empty().maxByOrNull(selector), null);
       expect(
-          [
-            [3]
-          ].maxByOrNull(selector),
-          [3]);
-      expect(iterable.maxByOrNull(selector), [1, 2, 3]);
+        Iterable.generate(
+          1,
+          (index) => Iterable<int>.generate(1),
+        ).maxByOrNull(selector),
+        [0],
+      );
+      expect(iterable.maxByOrNull(selector), [0, 1, 2]);
     });
 
     test('maxOf', () {
-      final iterable = [
-        [1],
-        [1, 2],
-        [1, 2, 3],
-      ];
-
-      Comparable selector(List element) => element.length;
+      final iterable = Iterable.generate(
+        3,
+        (index) => Iterable<int>.generate(index + 1),
+      );
+      Comparable selector(Iterable<int> element) => element.length;
 
       expect(
-        () => <List>[].maxOf(selector),
+        () => Iterable<Iterable<int>>.empty().maxOf(selector),
         throwsA(isA<NoSuchElementException>()),
       );
       expect(
-        [
-          [3]
-        ].maxOf(selector),
+        Iterable.generate(
+          1,
+          (index) => Iterable<int>.generate(1),
+        ).maxOf(selector),
         1,
       );
       expect(iterable.maxOf(selector), 3);
     });
 
     test('maxOfOrNull', () {
-      final iterable = [
-        [1],
-        [1, 2],
-        [1, 2, 3],
-      ];
+      final iterable = Iterable.generate(
+        3,
+        (index) => Iterable<int>.generate(index + 1),
+      );
+      Comparable selector(Iterable<int> element) => element.length;
 
-      Comparable selector(List element) => element.length;
-
-      expect(<List>[].maxOfOrNull(selector), null);
+      expect(Iterable<Iterable<int>>.empty().maxOfOrNull(selector), null);
       expect(
-        [
-          [3]
-        ].maxOfOrNull(selector),
+        Iterable.generate(
+          1,
+          (index) => Iterable<int>.generate(1),
+        ).maxOf(selector),
         1,
       );
-      expect(iterable.maxOfOrNull(selector), 3);
+      expect(iterable.maxOf(selector), 3);
     });
 
     test('maxOfWith', () {
@@ -430,10 +541,13 @@ void main() {
       int selector(int element) => -element;
 
       expect(
-        () => <int>[].maxOfWith(comparator, selector),
+        () => Iterable<int>.empty().maxOfWith(comparator, selector),
         throwsA(isA<NoSuchElementException>()),
       );
-      expect([3].maxOfWith(comparator, selector), -3);
+      expect(
+        Iterable.generate(1, (_) => 3).maxOfWith(comparator, selector),
+        -3,
+      );
       expect(iterable.maxOfWith(comparator, selector), 0);
     });
 
@@ -441,79 +555,81 @@ void main() {
       int comparator(int a, int b) => a.compareTo(b);
       int selector(int element) => -element;
 
-      expect(<int>[].maxOfWithOrNull(comparator, selector), null);
-      expect([3].maxOfWithOrNull(comparator, selector), -3);
+      expect(Iterable<int>.empty().maxOfWithOrNull(comparator, selector), null);
+      expect(
+        Iterable.generate(1, (_) => 3).maxOfWithOrNull(comparator, selector),
+        -3,
+      );
       expect(iterable.maxOfWithOrNull(comparator, selector), 0);
     });
 
     test('maxWithOrNull', () {
       int comparator(int a, int b) => a.compareTo(b);
 
-      expect(<int>[].maxWithOrNull(comparator), null);
-      expect([3].maxWithOrNull(comparator), 3);
+      expect(Iterable<int>.empty().maxWithOrNull(comparator), null);
+      expect(Iterable.generate(1, (_) => 3).maxWithOrNull(comparator), 3);
       expect(iterable.maxWithOrNull(comparator), 2);
     });
   });
 
   group('Min', () {
     test('minOrNull', () {
-      expect(<num>[].minOrNull, null);
+      expect(empty.minOrNull, null);
       expect(iterable.minOrNull, 0);
     });
 
     test('minByOrNull', () {
-      final iterable = [
-        [1],
-        [1, 2],
-        [1, 2, 3],
-      ];
-      Comparable selector(List element) => element.length;
+      final iterable = Iterable.generate(
+        3,
+        (index) => Iterable<int>.generate(index + 1),
+      );
+      Comparable selector(Iterable<int> element) => element.length;
 
-      expect(<List>[].minByOrNull(selector), null);
+      expect(Iterable<Iterable<int>>.empty().minByOrNull(selector), null);
       expect(
-          [
-            [3]
-          ].minByOrNull(selector),
-          [3]);
-      expect(iterable.minByOrNull(selector), [1]);
+        Iterable.generate(
+          1,
+          (index) => Iterable<int>.generate(1),
+        ).minByOrNull(selector),
+        [0],
+      );
+      expect(iterable.minByOrNull(selector), [0]);
     });
 
     test('minOf', () {
-      final iterable = [
-        [1],
-        [1, 2],
-        [1, 2, 3],
-      ];
-
-      Comparable selector(List element) => element.length;
+      final iterable = Iterable.generate(
+        3,
+        (index) => Iterable<int>.generate(index + 1),
+      );
+      Comparable selector(Iterable<int> element) => element.length;
 
       expect(
-        () => <List>[].minOf(selector),
+        () => Iterable<Iterable<int>>.empty().minOf(selector),
         throwsA(isA<NoSuchElementException>()),
       );
       expect(
-        [
-          [3]
-        ].minOf(selector),
+        Iterable.generate(
+          1,
+          (index) => Iterable<int>.generate(1),
+        ).minOf(selector),
         1,
       );
       expect(iterable.minOf(selector), 1);
     });
 
     test('minOfOrNull', () {
-      final iterable = [
-        [1],
-        [1, 2],
-        [1, 2, 3],
-      ];
+      final iterable = Iterable.generate(
+        3,
+        (index) => Iterable<int>.generate(index + 1),
+      );
+      Comparable selector(Iterable<int> element) => element.length;
 
-      Comparable selector(List element) => element.length;
-
-      expect(<List>[].minOfOrNull(selector), null);
+      expect(Iterable<Iterable<int>>.empty().minOfOrNull(selector), null);
       expect(
-        [
-          [3]
-        ].minOfOrNull(selector),
+        Iterable.generate(
+          1,
+          (index) => Iterable<int>.generate(1),
+        ).minOfOrNull(selector),
         1,
       );
       expect(iterable.minOfOrNull(selector), 1);
@@ -524,10 +640,13 @@ void main() {
       int selector(int element) => -element;
 
       expect(
-        () => <int>[].minOfWith(comparator, selector),
+        () => Iterable<int>.empty().minOfWith(comparator, selector),
         throwsA(isA<NoSuchElementException>()),
       );
-      expect([3].minOfWith(comparator, selector), -3);
+      expect(
+        Iterable.generate(1, (index) => 3).minOfWith(comparator, selector),
+        -3,
+      );
       expect(iterable.minOfWith(comparator, selector), -2);
     });
 
@@ -535,24 +654,36 @@ void main() {
       int comparator(int a, int b) => a.compareTo(b);
       int selector(int element) => -element;
 
-      expect(<int>[].minOfWithOrNull(comparator, selector), null);
-      expect([3].minOfWithOrNull(comparator, selector), -3);
+      expect(Iterable<int>.empty().minOfWithOrNull(comparator, selector), null);
+      expect(
+        Iterable.generate(
+          1,
+          (index) => 3,
+        ).minOfWithOrNull(comparator, selector),
+        -3,
+      );
       expect(iterable.minOfWithOrNull(comparator, selector), -2);
     });
 
     test('minWithOrNull', () {
       int comparator(int a, int b) => a.compareTo(b);
 
-      expect(<int>[].minWithOrNull(comparator), null);
-      expect([3].minWithOrNull(comparator), 3);
+      expect(Iterable<int>.empty().minWithOrNull(comparator), null);
+      expect(
+        Iterable.generate(
+          1,
+          (index) => 3,
+        ).minWithOrNull(comparator),
+        3,
+      );
       expect(iterable.minWithOrNull(comparator), 0);
     });
   });
 
   test('none', () {
-    expect([].none(), true);
+    expect(empty.none(), true);
     expect(iterable.none(), false);
-    expect([].none((element) => element == null), true);
+    expect(Iterable.empty().none((element) => element == null), true);
     expect(iterable.none((element) => element.isNegative), true);
     expect(iterable.none((element) => element.isEven), false);
   });
@@ -560,26 +691,32 @@ void main() {
   group('OnEach', () {
     test('onEach', () {
       var idx = 0;
-      final iter = iterable.onEach((element) {
-        expect(element, idx++);
-      });
-      expect(iter, iterable);
+      void action(int element) => expect(element, idx++);
+
+      expect(empty.onEach(action), empty);
+      expect(iterable.onEach(action), iterable);
     });
 
     test('onEachIndexed', () {
       var idx = 0;
-      final iter = iterable.onEachIndexed((index, element) {
+      void action(int index, int element) {
         expect(element, idx);
         expect(index, idx);
         idx++;
-      });
-      expect(iter, iterable);
+      }
+
+      expect(empty.onEachIndexed(action), empty);
+      expect(iterable.onEachIndexed(action), iterable);
     });
   });
 
   test('partition', () {
-    final partition = iterable.partition((element) => element.isEven);
+    bool test(int element) => element.isEven;
+    final emptyPartition = empty.partition(test);
+    final partition = iterable.partition(test);
 
+    expect(emptyPartition.first, []);
+    expect(emptyPartition.second, []);
     expect(partition.first, [0, 2]);
     expect(partition.second, [1]);
   });
@@ -589,7 +726,7 @@ void main() {
       int combine(int index, int value, int element) => index + value + element;
 
       expect(
-        () => <int>[].reduceIndexed(combine),
+        () => empty.reduceIndexed(combine),
         throwsA(isA<NoSuchElementException>()),
       );
       expect(iterable.reduceIndexed(combine), 6);
@@ -598,21 +735,21 @@ void main() {
     test('reduceIndexedOrNull', () {
       int combine(int index, int value, int element) => index + value + element;
 
-      expect(<int>[].reduceIndexedOrNull(combine), null);
+      expect(empty.reduceIndexedOrNull(combine), null);
       expect(iterable.reduceIndexedOrNull(combine), 6);
     });
 
     test('reduceOrNull', () {
       int combine(int value, int element) => value + element;
 
-      expect(<int>[].reduceOrNull(combine), null);
+      expect(empty.reduceOrNull(combine), null);
       expect(iterable.reduceOrNull(combine), 3);
     });
 
     test('runningReduce', () {
       int combine(previousValue, element) => previousValue + element;
 
-      expect([].runningReduce(combine), []);
+      expect(empty.runningReduce(combine), []);
       expect(iterable.runningReduce(combine), [1, 3]);
     });
 
@@ -620,15 +757,18 @@ void main() {
       int combine(index, previousValue, element) =>
           previousValue + element + index;
 
-      expect([].runningReduceIndexed(combine), []);
+      expect(empty.runningReduceIndexed(combine), []);
       expect(iterable.runningReduceIndexed(combine), [2, 6]);
     });
   });
 
   test('requireNoNulls', () {
-    expect(<int?>[0, 1, 2].requireNoNulls, [0, 1, 2]);
+    expect(Iterable<int?>.generate(3).requireNoNulls, [0, 1, 2]);
     expect(
-      () => <int?>[0, null, 2].requireNoNulls,
+      () => Iterable<int?>.generate(
+        3,
+        (index) => index.isOdd ? null : index,
+      ).requireNoNulls,
       throwsA(isArgumentError),
     );
   });
@@ -639,108 +779,136 @@ void main() {
 
   group('Single', () {
     test('singleOrNull', () {
-      expect(<int>[].singleOrNull, null);
-      expect([0].singleOrNull, 0);
-      expect([0, 1].singleOrNull, null);
+      expect(empty.singleOrNull, null);
+      expect(Iterable<int>.generate(1).singleOrNull, 0);
+      expect(Iterable<int>.generate(2).singleOrNull, null);
     });
 
     test('singleWhrerOrNull', () {
       bool test(int element) => element == 1;
 
-      expect(<int>[].singleWhereOrNull(test), null);
-      expect([0, 2].singleWhereOrNull(test), null);
-      expect([0, 1, 2].singleWhereOrNull(test), 1);
-      expect([0, 1, 2, 1].singleWhereOrNull(test), null);
+      expect(empty.singleWhereOrNull(test), null);
+      expect(
+        Iterable<int>.generate(2, (index) => index * 2).singleWhereOrNull(test),
+        null,
+      );
+      expect(Iterable<int>.generate(3).singleWhereOrNull(test), 1);
+      expect(
+        Iterable<int>.generate(2, (index) => 1).singleWhereOrNull(test),
+        null,
+      );
     });
   });
 
   group('Sort', () {
     test('sorted', () {
-      expect([2, 1, 0].sorted, [0, 1, 2]);
+      expect(reversedIterable.sorted, [0, 1, 2]);
     });
 
     test('sortedBy', () {
-      expect([0, 1, 2].sortedBy((element) => -element), [2, 1, 0]);
+      expect(iterable.sortedBy((element) => -element), [2, 1, 0]);
     });
 
     test('sortedByDescending', () {
-      expect([0, 1, 2].sortedByDescending((element) => -element), [0, 1, 2]);
+      expect(iterable.sortedByDescending((element) => -element), [0, 1, 2]);
     });
 
     test('sortedDescending', () {
-      expect([0, 1, 2].sortedDescending, [2, 1, 0]);
+      expect(iterable.sortedDescending, [2, 1, 0]);
     });
 
     test('sortedWith', () {
-      expect([2, 1, 0].sortedWith((a, b) => a.compareTo(b)), [0, 1, 2]);
-      expect([0, 1, 2].sortedWith((a, b) => b.compareTo(a)), [2, 1, 0]);
+      expect(
+        reversedIterable.sortedWith((a, b) => a.compareTo(b)),
+        [0, 1, 2],
+      );
+      expect(iterable.sortedWith((a, b) => b.compareTo(a)), [2, 1, 0]);
     });
   });
 
   group('Sum', () {
+    final iterable = Iterable<num>.generate(
+      3,
+      (index) => index.isOdd ? index.toDouble() : index,
+    );
     test('sum', () {
-      expect(<num>[].sum, 0);
-      expect([0, 2.0, -1].sum, 1.0);
+      expect(empty.sum, 0);
+      expect(iterable.sum, 3);
     });
 
     test('sumOf', () {
-      expect(<num>[].sumOf((element) => -element), 0);
-      expect([0, 2.0, -1].sumOf((element) => -element), -1.0);
+      num selector(num element) => -element;
+
+      expect(empty.sumOf(selector), 0);
+      expect(iterable.sumOf(selector), -3.0);
     });
   });
 
   test('union', () {
-    expect([].union([1]), [1]);
-    expect([1].union([1]), [1]);
-    expect([1].union([1, 2]), [1, 2]);
-    expect([1, 2].union([2, 3]), [1, 2, 3]);
-    expect([1, 2].union([3, 4]), [1, 2, 3, 4]);
-    expect([1, 2, 2].union([]), [1, 2]);
-    expect([1, 4, 2].union([3, 2]), [1, 4, 2, 3]);
+    expect(empty.union([1]), [1]);
+    expect(Iterable.generate(1).union([0]), [0]);
+    expect(Iterable.generate(1).union([0, 1]), [0, 1]);
+    expect(Iterable.generate(2).union([1, 2]), [0, 1, 2]);
+    expect(Iterable.generate(2).union([2, 3]), [0, 1, 2, 3]);
+    expect(
+      Iterable.generate(
+        3,
+        (index) => index.isEven ? 0 : index,
+      ).union([]),
+      [0, 1],
+    );
+    expect(
+      Iterable.generate(
+        3,
+        (index) => index.isEven ? 0 : index,
+      ).union([1, 0]),
+      [0, 1],
+    );
   });
 
   test('unzip', () {
-    final unzip = [
-      Pair(0, -0),
-      Pair(1, -1),
-      Pair(2, -2),
-    ].unzip();
+    final emptyUnzip = Iterable<Pair>.empty().unzip();
+    final unzip = Iterable.generate(3, (index) => Pair(index, -index)).unzip();
 
-    expect(
-      unzip.first,
-      Iterable<int>.generate(3),
-    );
-    expect(
-      unzip.second,
-      Iterable<int>.generate(3, (index) => -index),
-    );
+    expect(emptyUnzip.first, Iterable<int>.empty());
+    expect(emptyUnzip.second, Iterable<int>.empty());
+    expect(unzip.first, Iterable<int>.generate(3));
+    expect(unzip.second, Iterable<int>.generate(3, (index) => -index));
   });
 
   group('Where', () {
     test('whereIndexed', () {
-      final whereIndexed = iterable.whereIndexed(
-        (index, element) => (index + element).isEven,
-      );
+      bool test(int index, int element) => (index + element).isEven;
 
-      expect(whereIndexed, [0, 1, 2]);
+      expect(empty.whereIndexed(test), []);
+      expect(iterable.whereIndexed(test), [0, 1, 2]);
     });
 
     test('whereIsInstance', () {
-      final whereIsInstance = [0, '1', 2].whereIsInstance<String>();
+      final iterable = Iterable.generate(
+        3,
+        (index) => index.isOdd ? index.toString() : index,
+      );
 
-      expect(whereIsInstance, ['1']);
+      expect(empty.whereIsInstance<String>(), []);
+      expect(iterable.whereIsInstance<String>(), ['1']);
     });
 
     test('whereNot', () {
-      final whereNot = iterable.whereNot((element) => element.isEven);
+      bool test(int element) => element.isEven;
 
-      expect(whereNot, [1]);
+      expect(empty.whereNot(test), []);
+      expect(iterable.whereNot(test), [1]);
     });
 
     test('whereNotNull', () {
-      final whereNotNull = [null, 1, null, 3].whereNotNull();
+      final iterable = Iterable.generate(
+        4,
+        (index) => index.isEven ? null : index,
+      );
 
-      expect(whereNotNull, [1, 3]);
+      expect(empty.whereNotNull, []);
+      expect(iterable.whereNotNull, [1, 3]);
     });
   });
 
@@ -748,6 +916,7 @@ void main() {
     test('windowed', () {
       final iterable = Iterable.generate(10);
 
+      expect(empty.windowed(size: 1), []);
       expect(
         () => iterable.windowed(size: -1, step: -1),
         throwsArgumentError,
@@ -799,8 +968,10 @@ void main() {
 
     test('windowedAndTransform', () {
       final iterable = Iterable<int>.generate(10);
-      String transform(int element) => element.toString();
+      int transform(Iterable<int> chunk) =>
+          chunk.reduce((value, element) => value + element);
 
+      expect(empty.windowedAndTransform(size: 1, transform: transform), []);
       expect(
         () => iterable.windowedAndTransform(
           size: -1,
@@ -834,18 +1005,8 @@ void main() {
         throwsArgumentError,
       );
       expect(
-        iterable.windowedAndTransform(
-          size: 5,
-          transform: transform,
-        ),
-        [
-          ['0', '1', '2', '3', '4'],
-          ['1', '2', '3', '4', '5'],
-          ['2', '3', '4', '5', '6'],
-          ['3', '4', '5', '6', '7'],
-          ['4', '5', '6', '7', '8'],
-          ['5', '6', '7', '8', '9'],
-        ],
+        iterable.windowedAndTransform(size: 5, transform: transform),
+        [10, 15, 20, 25, 30, 35],
       );
       expect(
         iterable.windowedAndTransform(
@@ -853,10 +1014,7 @@ void main() {
           step: 3,
           transform: transform,
         ),
-        [
-          ['0', '1', '2', '3', '4'],
-          ['3', '4', '5', '6', '7'],
-        ],
+        [10, 25],
       );
       expect(
         iterable.windowedAndTransform(
@@ -865,12 +1023,7 @@ void main() {
           transform: transform,
           partialWindows: true,
         ),
-        [
-          ['0', '1', '2', '3', '4'],
-          ['3', '4', '5', '6', '7'],
-          ['6', '7', '8', '9'],
-          ['9'],
-        ],
+        [10, 25, 30, 9],
       );
     });
   });
@@ -885,27 +1038,24 @@ void main() {
 
   group('Zip', () {
     test('zip', () {
-      expect([1, 2].zip(['1', '2', '3']), [Pair(1, '1'), Pair(2, '2')]);
+      expect(
+        iterable.zip(['0', '1', '2', '3']),
+        [Pair(0, '0'), Pair(1, '1'), Pair(2, '2')],
+      );
     });
 
     test('zipAndTransform', () {
       expect(
-        [1, 2].zipAndTransform(
-          ['1', '2', '3'],
+        iterable.zipAndTransform(
+          ['0', '1', '2', '3'],
           (pair) => '${pair.first}${pair.second}',
         ),
-        ['11', '22'],
+        ['00', '11', '22'],
       );
     });
 
     test('zipWithNext', () {
-      expect(
-        iterable.zipWithNext,
-        [
-          Pair(0, 1),
-          Pair(1, 2),
-        ],
-      );
+      expect(iterable.zipWithNext, [Pair(0, 1), Pair(1, 2)]);
     });
 
     test('zipWithNextAndTransform', () {
@@ -920,28 +1070,36 @@ void main() {
 
   group('Operator', () {
     test('-', () {
-      expect([] - [1], []);
-      expect([1] - [], [1]);
-      expect([1] - [1], []);
-      expect([1] - [1, 2], []);
-      expect([1, 2] - [1], [2]);
-      expect([1, 2] - [1, 3], [2]);
-      expect([1, 2] - [3, 4], [1, 2]);
-      expect([1, 1, 2] - [1], [2]);
-      expect([1, 1, 2] - [1, 2], []);
+      expect(empty - [0], []);
+      expect(Iterable.generate(1) - [], [0]);
+      expect(Iterable.generate(1) - [0], []);
+      expect(Iterable.generate(1) - [0, 1], []);
+      expect(Iterable.generate(2) - [0], [1]);
+      expect(Iterable.generate(2) - [0, 2], [1]);
+      expect(Iterable.generate(2) - [2, 3], [0, 1]);
+      expect(
+        Iterable.generate(3, (index) => index.isOdd ? index - 1 : index) - [0],
+        [2],
+      );
+      expect(
+        Iterable.generate(3, (index) => index.isOdd ? index - 1 : index) -
+            [0, 2],
+        [],
+      );
     });
 
     test('+', () {
-      expect([] + [2], [2]);
-      expect([1] + [], [1]);
-      expect([1] + [2], [1, 2]);
-      expect([2] + [1], [2, 1]);
-      expect([1] + [1, 2], [1, 1, 2]);
-      expect([1, 2] + [1], [1, 2, 1]);
-      expect([1, 2] + [3, 4], [1, 2, 3, 4]);
+      expect(empty + [0], [0]);
+      expect(Iterable.generate(1) + [], [0]);
+      expect(Iterable.generate(1) + [1], [0, 1]);
+      expect(Iterable.generate(1, (index) => index + 1) + [0], [1, 0]);
+      expect(Iterable.generate(1) + [0, 1], [0, 0, 1]);
+      expect(Iterable.generate(2) + [0], [0, 1, 0]);
+      expect(Iterable.generate(2) + [2, 3], [0, 1, 2, 3]);
     });
 
     test('<<', () {
+      expect(empty << 1, []);
       expect(() => iterable << -1, throwsA(isArgumentError));
       expect(iterable << 0, [0, 1, 2]);
       expect(iterable << 1, [1, 2]);
@@ -951,6 +1109,7 @@ void main() {
     });
 
     test('>>', () {
+      expect(() => empty >> 1, throwsA(isArgumentError));
       expect(() => iterable >> -1, throwsA(isArgumentError));
       expect(iterable >> 0, [0, 1, 2]);
       expect(iterable >> 1, [0, 1]);
@@ -960,12 +1119,19 @@ void main() {
     });
 
     test('^', () {
-      expect([] ^ [], []);
-      expect([1] ^ [], [1]);
-      expect([1] ^ [2], [1, 2]);
-      expect([1, 1] ^ [2], [1, 2]);
-      expect([1, 1] ^ [2, 2], [1, 2]);
-      expect([1, 1, 2] ^ [2, 2, 3], [1, 3]);
+      expect(empty ^ [], []);
+      expect(Iterable.generate(1) ^ [], [0]);
+      expect(Iterable.generate(1) ^ [1], [0, 1]);
+      expect(Iterable.generate(2, (_) => 0) ^ [1], [0, 1]);
+      expect(Iterable.generate(2, (_) => 0) ^ [1, 1], [0, 1]);
+      expect(
+        Iterable.generate(
+              3,
+              (index) => index.isOdd ? index - 1 : index,
+            ) ^
+            [2, 2, 3],
+        [0, 3],
+      );
     });
   });
 }